requirements.txt
numpy>=1.18.5
opencv-python>=4.1.2
torch>=1.7.0
torchvision>=0.8.1
matplotlib>=3.2.2
PyYAML>=5.3.1
tqdm>=4.41.0
pillow>=7.1.2
scipy>=1.4.1
pandas>=1.1.4
seaborn>=0.11.0
mysql-connector-python>=8.0.24
pyserial>=3.5
thop>=0.1.1  # FLOPs计算
